Database/Final Project/requirements.txt
streamlit
google-generativeai
python-dotenv
mysqldump
regex
mysql-connector-python
